fix(email-notifier):  migrate constants to separate file, add misc dir
