+++ integrations/email-notifier/src/misc/constants.ts
@@ -1,0 +1,2 @@
+export const FROM_EMAIL_ADDRESS = "noreply@bp-mailer.com";
+export const CONTACT_LIST = "default";
